+++ README.template.md
@@ -44,11 +44,11 @@ cd <your project name>
 
 - 画面表示において複数画面内で共通して使用するユーティリティなどを配置する。
 - :ui モジュール配下にサブモジュール、ソースコードを配置する。
-- 各画面の UI 実装は :ui モジュール配下ではなく :feature モジュール配下に配置する。
+- 各画面の UI 実装は :ui:feature モジュール配下に配置する。
 
 ### 各機能
 
-- 各機能の UI は :feature 以下にモジュールを切って配置する。
+- 各機能の UI は :ui:feature 以下にモジュールを切って配置する。
 - 各 feature モジュールは `libs.plugins.buildLogicModuleFeature` Gradle Plugin を適用する。
 - モジュールは機能グループごとにきり、モジュール内にパッケージを切って各画面を配置する。
 - 機能グループモジュールごとに Navigation.kt を配置し、グループ内での画面遷移のロジックを配置する。
