+++ app/build.gradle.kts
@@ -25,10 +25,11 @@ android {
 }
 
 dependencies {
-    // feature modules
     implementation(projects.data)
     implementation(projects.domain)
     implementation(projects.ui)
     implementation(projects.debug)
-    implementation(projects.feature.example)
+
+    // feature modules
+    implementation(projects.ui.feature.example)
 }
