--- feature/example/build.gradle.kts
@@ -1,10 +1,0 @@
-plugins {
-    alias(libs.plugins.buildLogicModuleFeature)
-}
-
-android {
-    namespace = "${libs.versions.app.applicationId.get()}.feature.example"
-}
-
-dependencies {
-}
