--- feature/example/src/main/kotlin/your/projectPackage/feature/example/apiPostList/ExampleApiPostListUiAction.kt
@@ -1,5 +1,0 @@
-package your.projectPackage.feature.example.apiPostList
-
-internal sealed interface ExampleApiPostListUiAction {
-    data object Refresh : ExampleApiPostListUiAction
-}
