--- feature/example/src/main/kotlin/your/projectPackage/feature/example/counter/ExampleCounterUiAction.kt
@@ -1,5 +1,0 @@
-package your.projectPackage.feature.example.counter
-
-internal sealed interface ExampleCounterUiAction {
-    data object Refresh : ExampleCounterUiAction
-}
