--- feature/example/src/main/kotlin/your/projectPackage/feature/example/counter/ExampleCounterUiState.kt
@@ -1,5 +1,0 @@
-package your.projectPackage.feature.example.counter
-
-internal data class ExampleCounterUiState(
-    val count: Int,
-)
