--- feature/example/src/main/kotlin/your/projectPackage/feature/example/localDbUserList/ExampleLocalDbUserListUiAction.kt
@@ -1,8 +1,0 @@
-package your.projectPackage.feature.example.localDbUserList
-
-import your.projectPackage.domain.example.user.User
-
-sealed interface ExampleLocalDbUserListUiAction {
-    data object OnAddUser : ExampleLocalDbUserListUiAction
-    data class OnDeleteUser(val user: User) : ExampleLocalDbUserListUiAction
-}
