--- feature/example/src/main/kotlin/your/projectPackage/feature/example/localDbUserList/ExampleLocalDbUserListUiState.kt
@@ -1,10 +1,0 @@
-package your.projectPackage.feature.example.localDbUserList
-
-import your.projectPackage.domain.example.user.User
-
-internal sealed interface ExampleLocalDbUserListUiState {
-    data object InitialLoading : ExampleLocalDbUserListUiState
-    data class Success(
-        val users: List<User>,
-    ) : ExampleLocalDbUserListUiState
-}
