+++ settings.gradle.kts
@@ -37,5 +37,5 @@ include(":common")
 include(":debug")
 include(":testing")
 include(":lintRules:composeNavigation")
-include(":feature:example")
 include(":ui:testing")
+include(":ui:feature:example")
