+++ ui/feature/example/build.gradle.kts
@@ -1,0 +1,10 @@
+plugins {
+    alias(libs.plugins.buildLogicModuleFeature)
+}
+
+android {
+    namespace = "${libs.versions.app.applicationId.get()}.ui.feature.example"
+}
+
+dependencies {
+}
