+++ ui/feature/example/src/main/AndroidManifest.xml
