+++ ui/feature/example/src/main/kotlin/your/projectPackage/ui/feature/example/apiPostList/ExampleApiPostListUiAction.kt
@@ -1,0 +1,5 @@
+package your.projectPackage.ui.feature.example.apiPostList
+
+internal sealed interface ExampleApiPostListUiAction {
+    data object Refresh : ExampleApiPostListUiAction
+}
