+++ ui/feature/example/src/main/kotlin/your/projectPackage/ui/feature/example/apiPostList/component/ErrorSection.kt
@@ -1,0 +1,42 @@
+package your.projectPackage.ui.feature.example.apiPostList.component
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import your.projectPackage.ui.PreviewRoot
+
+@Composable
+internal fun ErrorSection(
+    message: String,
+    modifier: Modifier = Modifier,
+) {
+    Surface(
+        modifier = modifier,
+        shape = MaterialTheme.shapes.medium,
+        color = MaterialTheme.colorScheme.errorContainer,
+        contentColor = MaterialTheme.colorScheme.error,
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(16.dp),
+        ) {
+            Text(message)
+        }
+    }
+}
+
+@Preview
+@Composable
+private fun ErrorSectionPreview() = PreviewRoot {
+    ErrorSection(
+        message = "通信エラー",
+    )
+}
