+++ ui/feature/example/src/main/kotlin/your/projectPackage/ui/feature/example/counter/ExampleCounterUiAction.kt
@@ -1,0 +1,5 @@
+package your.projectPackage.ui.feature.example.counter
+
+internal sealed interface ExampleCounterUiAction {
+    data object Refresh : ExampleCounterUiAction
+}
