+++ ui/feature/example/src/main/kotlin/your/projectPackage/ui/feature/example/counter/ExampleCounterUiState.kt
@@ -1,0 +1,5 @@
+package your.projectPackage.ui.feature.example.counter
+
+internal data class ExampleCounterUiState(
+    val count: Int,
+)
