+++ ui/feature/example/src/main/kotlin/your/projectPackage/ui/feature/example/localDbUserList/ExampleLocalDbUserListUiAction.kt
@@ -1,0 +1,8 @@
+package your.projectPackage.ui.feature.example.localDbUserList
+
+import your.projectPackage.domain.example.user.User
+
+sealed interface ExampleLocalDbUserListUiAction {
+    data object OnAddUser : ExampleLocalDbUserListUiAction
+    data class OnDeleteUser(val user: User) : ExampleLocalDbUserListUiAction
+}
