+++ ui/feature/example/src/main/kotlin/your/projectPackage/ui/feature/example/localDbUserList/ExampleLocalDbUserListUiState.kt
@@ -1,0 +1,10 @@
+package your.projectPackage.ui.feature.example.localDbUserList
+
+import your.projectPackage.domain.example.user.User
+
+internal sealed interface ExampleLocalDbUserListUiState {
+    data object InitialLoading : ExampleLocalDbUserListUiState
+    data class Success(
+        val users: List<User>,
+    ) : ExampleLocalDbUserListUiState
+}
